Conflicts:
	src/FOM/UserBundle/Controller/GroupController.php
	src/FOM/UserBundle/Controller/UserController.php
	src/FOM/UserBundle/Resources/views/Group/actions.html.twig
	src/FOM/UserBundle/Resources/views/Group/index.html.twig
	src/FOM/UserBundle/Resources/views/User/form.html.twig

+++ src/FOM/UserBundle/Controller/UserController.php
@@ -68,13 +68,17 @@ class UserController extends Controller {
             throw new AccessDeniedException();
         }
 
-        $form = $this->createForm(new UserType(), $user);
+        $profile = $this->addProfileForm($user);
+        $form = $this->createForm(new UserType(), $user, array(
+            'profile_formtype' => $profile['formtype']
+        ));
 
         return array(
             'user' => $user,
             'form' => $form->createView(),
             'form_name' => $form->getName(),
-            'edit' => false);
+            'edit' => false,
+            'profile_template' => $profile['template']);
     }
 
     /**
@@ -84,7 +88,10 @@ class UserController extends Controller {
      */
     public function createAction() {
         $user = new User();
-        $form = $this->createForm(new UserType(), $user);
+        $profile = $this->addProfileForm($user);
+        $form = $this->createForm(new UserType(), $user, array(
+            'profile_formtype' => $profile['formtype']
+        ));
 
 
 
@@ -115,6 +122,14 @@ class UserController extends Controller {
             $aclManager->setObjectACLFromForm($user, $form->get('acl'),
                 'object');
 
+            // Check and persists profile if exists
+            $profile = $user->getProfile();
+            if($profile) {
+                $em->persist($profile);
+            }
+
+            $em->flush();
+
             $this->get('session')->setFlash('success',
                 'The user has been saved.');
 
@@ -126,7 +141,8 @@ class UserController extends Controller {
             'user' => $user,
             'form' => $form->createView(),
             'form_name' => $form->getName(),
-            'edit' => false);
+            'edit' => false,
+            'profile_template' => $profile['template']);
     }
 
     /**
@@ -146,15 +162,18 @@ class UserController extends Controller {
             throw new AccessDeniedException();
         }
 
+        $profile = $this->addProfileForm($user);
         $form = $this->createForm(new UserType(), $user, array(
             'requirePassword' => false,
-            'extendedEdit' => $user->isAdmin('USER')));
+            'profile_formtype' => $profile['formtype']
+        ));
 
         return array(
             'user' => $user,
             'form' => $form->createView(),
             'form_name' => $form->getName(),
-            'edit' => true);
+            'edit' => true,
+            'profile_template' => $profile['template']);
     }
 
     /**
@@ -185,7 +204,11 @@ class UserController extends Controller {
             $keepPassword = true;
         }
 
-        $form = $this->createForm(new UserType(), $user);
+        $profile = $this->addProfileForm($user);
+        $form = $this->createForm(new UserType(), $user, array(
+            'requirePassword' => false,
+            'profile_formtype' => $profile['formtype']
+        ));
         $form->bind($userData);
 
         if($form->isValid()) {
@@ -202,6 +225,9 @@ class UserController extends Controller {
             $aclManager->setObjectACLFromForm($user, $form->get('acl'),
                 'object');
 
+            if($user->getProfile()) {
+                $em->persist($user->getProfile());
+            }
             $em->flush();
 
             $this->get('session')->setFlash('success', 'The user has been updated.');
@@ -214,7 +240,8 @@ class UserController extends Controller {
             'user' => $user,
             'form' => $form->createView(),
             'form_name' => $form->getName(),
-            'edit' => true);
+            'edit' => true,
+            'profile_template' => $profile['template']);
     }
 
     /**
@@ -235,14 +262,23 @@ class UserController extends Controller {
         }
 
         // ACL access check
-        // $securityContext = $this->get('security.context');
-        // if(false === $securityContext->isGranted('DELETE', $user)) {
-        //     throw new AccessDeniedException();
-        // }
+        $securityContext = $this->get('security.context');
+        if(false === $securityContext->isGranted('DELETE', $user)) {
+            throw new AccessDeniedException();
+        }
+
+        $this->addProfileForm($user);
+        $profile = $user->getProfile();
+        $form = $this->createDeleteForm($id);
+        $request = $this->getRequest();
 
         try {
             $em = $this->getDoctrine()->getEntityManager();
+
             $em->remove($user);
+            if($user->getProfile()) {
+                $em->remove($user->getProfile());
+            }
             $em->flush();
 
             $this->get('session')->setFlash('success',
@@ -252,6 +288,37 @@ class UserController extends Controller {
                 'The user couldn\'t be deleted.');
         }
         return new Response();
+    }
+
+    private function addProfileForm(User $user)
+    {
+        $container = $this->container;
+        $profileEntity = $container->getParameter('fom_user.profile_entity');
+        $profileFormType = $container->getParameter('fom_user.profile_formtype');
+        $profileTemplate = $container->getParameter('fom_user.profile_template');
+
+        if($profileEntity !== null && $profileFormType !== null) {
+            if($user->getId()) {
+                $profile = $this->getDoctrine()->getRepository($profileEntity)
+                    ->find($user->getId());
+                if(!$profile) {
+                $profile = new $profileEntity();
+                }
+            } else {
+                $profile = new $profileEntity();
+            }
+
+            $user->setProfile($profile);
+
+        } else {
+            $profileFormType = null;
+            $profileTemplate = null;
+        }
+
+        return array(
+            'formtype' => $profileFormType,
+            'template' => $profileTemplate
+        );
     }
 }
